🐜🍛 Checkpoint ./public/scripts/viewer.js:49008067/2323

+++ public/scripts/viewer.js
@@ -183,12 +183,21 @@ async function getData(tags, PID) {
 
     testImage(img);
     
+    
     // File URL Preload
-    const file = new Image();
-    file.src = tagData[i].file_url
-    file.remove()
+    if (settings.quality === "Full") {
+      if (!data.file_url.endsWith(".webm") || !data.file_url.endsWith(".mp4")) {
+        const file = new Image();
+        file.src = convertURL(tagData[i].file_url)
+        file.remove()
+      }
+    }
     
     // Sample URL Preload
-
+    if (settings.quality === "Sample") {
+        const sample = new Image();
+        sample.src = convertURL(tagData[i].sample_url);
+        sample.remove()
+    }
   }
 }
